+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/lists/001_Display_Mobs_In_Area.lua
@@ -14,7 +14,7 @@ command: ''
 packageName: ''
 ]]--
 
-local area = gmcp.Room.Info.area
+local area = ataxiaBasher_areaKey()
 if ataxiaBasher.targetList[area] then
 	ataxiaEcho("Displaying valid targets for this area.")
 	for i,v in pairs(ataxiaBasher.targetList[area]) do