+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/lists/004_Ignore_Mobs.lua
@@ -15,7 +15,7 @@ packageName: ''
 ]]--
 
 ataxiaBasher.mobIgnore = ataxiaBasher.mobIgnore or {}
-if ataxia.denizensHere and ataxiaBasher.targetList[gmcp.Room.Info.area] then
+if ataxia.denizensHere and ataxiaBasher.targetList[ataxiaBasher_areaKey()] then
 	ataxiaEcho("Displaying mobs currently in room, but not in ignore list.")
 	for i,v in pairs(ataxia.denizensHere) do
 		if not table.contains(ataxiaBasher.mobIgnore, v) then