+++ src_new/aliases/levi_ataxia/for_levi/levi_062424/lists/006_Dangerous_Mobs.lua
@@ -14,16 +14,16 @@ command: ''
 packageName: ''
 ]]--
 
-if ataxia.denizensHere and ataxiaBasher.targetList[gmcp.Room.Info.area] then
+if ataxia.denizensHere and ataxiaBasher.targetList[ataxiaBasher_areaKey()] then
 	if matches[1]:find("c") then
-		for _, mob in pairs(ataxiaBasher.targetList[gmcp.Room.Info.area]) do
+		for _, mob in pairs(ataxiaBasher.targetList[ataxiaBasher_areaKey()]) do
 			ataxiaBasher_dangerAdd(mob)
 		end
 		ataxiaEcho("Danger list has been updated to reflect the entire area.")
 	else
 		ataxiaEcho("Displaying mobs in the area. Click A/R to add/remove from the danger list.")
 	
-		for _, mob in pairs(ataxiaBasher.targetList[gmcp.Room.Info.area]) do
+		for _, mob in pairs(ataxiaBasher.targetList[ataxiaBasher_areaKey()]) do
 			echo("\n")
 			fg("red")
 			echoLink("[A]", [[ataxiaBasher_dangerAdd("]]..mob..[[")]], "Add "..mob.." to the danger list.", true)