+++ src_new/triggers/levi_ataxia/for_levi/leviticus/denizen_attacks_misc_lines/004_Denizen_Emotes.lua
@@ -41,7 +41,7 @@ patterns:
 
 local x = multimatches[1][1]:lower()
 
-for _, npc in pairs(ataxiaBasher.targetList[gmcp.Room.Info.area]) do
+for _, npc in pairs(ataxiaBasher.targetList[ataxiaBasher_areaKey()]) do
 	if string.find(x, npc:lower()) then
 		if not ataxiaBasher.manual then
 			deleteFull()